dspy/requirements.txt
openai
anthropic
python-dotenv
langchain
pydantic
dspy-ai
nltk
scikit-learn
numpy
pandas
PyMuPDF
fitz
pillow
langchain-openai
bs4